test: consolidate MockSigningSmartWalletKit across packages
