+++ services/ymax-planner/test/mocks.ts
@@ -453,89 +453,3 @@ export const createMockGmpExecutionEvent = (
     transactionHash: '0x1234567890abcdef1234567890abcdef12345678',
   };
 };
-
-type SubmitTxResponse = {
-  code: number;
-  height: number;
-  transactionHash: string;
-  sequence: number;
-};
-export class MockSigningSmartWalletKit {
-  #submittedTransactions: Array<{
-    method: string;
-    sequence: number;
-    timestamp: number;
-  }> = [];
-  #networkSequence: () => number;
-  #shouldSimulateSequenceConflicts = false;
-  #networkDelay = 20;
-  #mockTime = 1234567890000;
-  #sequenceCounter = 0;
-
-  constructor(getNetworkSequence: () => number, mockTime = 1234567890000) {
-    this.#networkSequence = getNetworkSequence;
-    this.#mockTime = mockTime;
-  }
-
-  enableSequenceConflictSimulation() {
-    this.#shouldSimulateSequenceConflicts = true;
-  }
-
-  async executeOffer(offer: any, signerData: any) {
-    return this.#submitTransaction('executeOffer', signerData);
-  }
-
-  async sendBridgeAction(action: any, fee: any, memo: any, signerData: any) {
-    return this.#submitTransaction('sendBridgeAction', signerData);
-  }
-
-  async #submitTransaction(
-    method: string,
-    signerData: any,
-  ): Promise<SubmitTxResponse> {
-    // Simulate network delay
-    await new Promise(resolve => setTimeout(resolve, this.#networkDelay));
-
-    const currentNetworkSequence = this.#networkSequence();
-
-    // Simulate sequence mismatch if enabled and sequence is out of sync
-    if (
-      this.#shouldSimulateSequenceConflicts &&
-      signerData.sequence < currentNetworkSequence
-    ) {
-      throw new Error(
-        `Broadcasting transaction failed with code 32 (codespace: sdk). Log: account sequence mismatch, expected ${currentNetworkSequence}, got ${signerData.sequence}: incorrect account sequence`,
-      );
-    }
-
-    // Record successful transaction
-    this.#submittedTransactions.push({
-      method,
-      sequence: signerData.sequence,
-      timestamp: this.#mockTime + this.#sequenceCounter++ * 1000, // Increment by 1 second for each transaction
-    });
-
-    return {
-      code: 0,
-      height: 3321450 + this.#submittedTransactions.length,
-      transactionHash: `hash_${method}_${signerData.sequence}`,
-      sequence: signerData.sequence,
-    };
-  }
-
-  getSubmittedTransactions() {
-    return this.#submittedTransactions;
-  }
-
-  clearTransactions() {
-    this.#submittedTransactions = [];
-  }
-
-  setNetworkDelay(delay: number) {
-    this.#networkDelay = delay;
-  }
-
-  setMockTime(time: number) {
-    this.#mockTime = time;
-  }
-}
